Add a few unit tests
I still haven't figured out how the script is supposed to work
