+++ editquality/bwds/tests/bad_edits_words.txt
@@ -1,0 +1,1 @@
+bad_edits_words
