+++ editquality/bwds/tests/no_docs.txt
@@ -1,0 +1,1 @@
+no_docs.txt
