+++ editquality/bwds/tests/words_db.txt
@@ -1,0 +1,1 @@
+words_db
